+++ src/Command/MigrateResetCommand.php
@@ -11,6 +11,7 @@
 
 namespace WouterJ\EloquentBundle\Command;
 
+use Illuminate\Console\OutputStyle;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -22,7 +23,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class MigrateResetCommand extends BaseMigrateCommand
 {
-    protected function configure()
+    protected function configure(): void
     {
         $this->setName('eloquent:migrate:reset')
             ->setDescription('Rollback all database migrations')
@@ -41,14 +42,19 @@ EOH
         ;
     }
 
-    protected function execute(InputInterface $i, OutputInterface $o)
+    protected function execute(InputInterface $i, OutputInterface $o): int
     {
         if (!$i->getOption('force') && !$this->askConfirmationInProd($i, $o)) {
-            return;
+            return 1;
         }
 
         $migrator = $this->getMigrator();
         $migrator->setConnection($i->getOption('database'));
+
+        $illuminateLte56 = method_exists($migrator, 'getNotes');
+        if (!$illuminateLte56) {
+            $migrator->setOutput(new OutputStyle($i, $o));
+        }
 
         if (!$migrator->repositoryExists()) {
             $o->writeln('<error>Migration table not found.</>');
@@ -58,8 +64,12 @@ EOH
 
         $migrator->reset($this->getMigrationPaths($i), $i->getOption('pretend'));
 
-        foreach ($migrator->getNotes() as $note) {
-            $o->writeln($note);
+        if ($illuminateLte56) {
+            foreach ($migrator->getNotes() as $note) {
+                $o->writeln($note);
+            }
         }
+
+        return 0;
     }
 }