+++ tests/Command/MigrateRollbackCommandTest.php
@@ -11,6 +11,8 @@
 
 namespace WouterJ\EloquentBundle\Command;
 
+use Illuminate\Console\OutputStyle;
+use Symfony\Bridge\PhpUnit\SetUpTearDownTrait;
 use Symfony\Component\DependencyInjection\Container;
 use WouterJ\EloquentBundle\Migrations\Migrator;
 use PHPUnit\Framework\TestCase;
@@ -21,13 +23,20 @@ use Prophecy\Argument;
  */
 class MigrateRollbackCommandTest extends TestCase
 {
+    use SetUpTearDownTrait;
+
     private $command;
     private $migrator;
 
-    protected function setUp()
+    protected function doSetUp()
     {
         $this->migrator = $this->prophesize(Migrator::class);
-        $this->migrator->getNotes()->willReturn([]);
+        if (method_exists(Migrator::class, 'getNotes')) {
+            $this->migrator->getNotes()->willReturn([]);
+        } else {
+            $this->migrator->setOutput(Argument::type(OutputStyle::class))->willReturn();
+        }
+
         $this->migrator->paths()->willReturn([]);
         $this->migrator->setConnection(Argument::any())->willReturn();
 
@@ -125,6 +134,10 @@ class MigrateRollbackCommandTest extends TestCase
     /** @test */
     public function it_outputs_migration_notes()
     {
+        if (!method_exists(Migrator::class, 'getNotes')) {
+            $this->markTestSkipped('Only applies to Illuminate <5.7');
+        }
+
         $this->migrator->getNotes()->willReturn([
             'Rolled back: CreateFlightsTable',
             'Rolled back: SomethingToTest',