+++ tests/Migrations/CreatorTest.php
@@ -13,13 +13,16 @@ namespace WouterJ\EloquentBundle\Migrations;
 
 use Illuminate\Database\Schema\Blueprint;
 use PHPUnit\Framework\TestCase;
+use Symfony\Bridge\PhpUnit\SetUpTearDownTrait;
 
 class CreatorTest extends TestCase
 {
+    use SetUpTearDownTrait;
+
     protected $subject;
     protected $migrationsPath;
 
-    protected function setUp()
+    protected function doSetUp()
     {
         $this->subject = new Creator();
         $this->migrationsPath = sys_get_temp_dir();
@@ -61,8 +64,11 @@ class CreatorTest extends TestCase
 
         switch ($name) {
             case 'create':
-                $version = method_exists(Blueprint::class, 'dropSoftDeletesTz') ? 'new' : 'old';
-                $expected = $normalize(file_get_contents(__DIR__.'/../Fixtures/migrations/'.$name.'-'.$version.'.php'));
+                if (method_exists(Blueprint::class, 'integerIncrements')) {
+                    $name .= '-5.8';
+                }
+
+                $expected = $normalize(file_get_contents(__DIR__.'/../Fixtures/migrations/'.$name.'.php'));
 
                 break;
             default: